+++ lib/models/expense.dart
@@ -1,6 +1,6 @@
 class Expense {
   final String category;
-  final int amount;
+  final double amount;
   final DateTime datetime;
   final String description;
   final String id;
@@ -9,7 +9,7 @@ class Expense {
   const Expense(this.id, this.amount, this.category, this.description, this.datetime, this.prompt);
 
   factory Expense.fromJson(json) {
-    return Expense(json['_id'] as String, json['amount'] as int, json['category'] as String,
+    return Expense(json['_id'] as String, json['amount'] as double, json['category'] as String,
         json['description'] as String, DateTime.parse(json['date']).toLocal(), json['prompt'] ?? "");
   }
